+++ tests/util/run-with-preload.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+. "$(dirname -- "$0")"/preload.sh
+
+cd tests || exit 1
+
+preload_setup
+"$@"
+preload_cleanup